Refactoring && adding internationalization
