+++ mobile/app/build.gradle.kts
@@ -48,10 +48,7 @@ dependencies {
 
     // ZXing for QR code generation and scanning
     implementation("com.journeyapps:zxing-android-embedded:4.3.0")
-    implementation("com.google.zxing:core:3.4.1")
-
-    // ML Kit for Barcode Scanning
-    implementation("com.google.mlkit:barcode-scanning:17.0.3")
+    implementation("com.google.zxing:core:3.5.3")
 
     implementation("com.squareup.retrofit2:retrofit:2.9.0")
     implementation("com.squareup.retrofit2:converter-gson:2.9.0")
