+++ mobile/app/src/main/AndroidManifest.xml
@@ -19,16 +19,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MHealthy"
         tools:targetApi="31">
+
         <activity
-            android:name=".MainActivity"
-            android:exported="false" />
-        <activity
-            android:name=".QRScanActivity"
-            android:exported="false"
-            android:stateNotNeeded="true"
-            android:theme="@style/Theme.AppCompat.NoActionBar"
-            android:windowSoftInputMode="stateAlwaysHidden" /> <!-- <activity android:name=".QRScanActivity" -->
-        <!-- android:exported="false"/> -->
+            android:name=".ui.QRScanActivity"
+            android:screenOrientation="fullSensor"
+            tools:replace="screenOrientation" />
         <activity
             android:name=".AuthenticationActivity"
             android:exported="true">
@@ -39,18 +34,8 @@
             </intent-filter>
         </activity>
 
-<!--        <activity-->
-<!--            android:name=".MainActivity"-->
-<!--            android:exported="true">-->
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-<!--        </activity>-->
-
         <activity
-            android:name=".PatientMainActivity"
+            android:name=".ui.PatientMainActivity"
             android:exported="false" />
     </application>
