+++ mobile/app/src/main/res/layout/activity_authentication.xml
@@ -17,13 +17,13 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        android:contentDescription="App Logo" />
+        android:contentDescription="@string/app_name" />
 
     <TextView
         android:id="@+id/tvTitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Authentication"
+        android:text="@string/authenticate"
         android:textSize="24sp"
         android:textStyle="bold"
         android:textColor="@color/text_color"
@@ -60,7 +60,7 @@
                     android:id="@+id/etIdToken"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:hint="ID or Token"
+                    android:hint="@string/token"
                     android:inputType="text"
                     android:minHeight="48dp" />
             </com.google.android.material.textfield.TextInputLayout>
@@ -69,14 +69,14 @@
                 android:id="@+id/btnLogin"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="Login"
+                android:text="@string/login"
                 android:layout_marginTop="24dp"
                 android:minHeight="56dp" />
 
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="OR"
+                android:text="@string/or"
                 android:textSize="18sp"
                 android:layout_gravity="center"
                 android:layout_marginTop="16dp"
@@ -86,7 +86,7 @@
                 android:id="@+id/btnQrScan"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="Scan QR Code"
+                android:text="@string/scan_qr"
                 app:icon="@drawable/ic_qr_code"
                 app:iconGravity="textStart"
                 android:minHeight="56dp" />
